starter-backend/requirements.txt
fastapi==0.128.0
uvicorn==0.40.0
pydantic==2.10.3
python-dotenv==1.1.0
asyncpg
httpx==0.28.1
pytest==8.3.4
